pyproject.toml
[project]
name = "sub2clash"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "pycryptodome>=3.23.0",
    "requests>=2.33.1",
]
